Added custom state class
